+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- move `topological_sort` from extractors to common
 - better align mex.common.models with mex-model
 - update mex-model dependency to 4.3
 - parse_csv now summarizes error logs in batches with amount of successful validations.